Refactor: Extract post_login and auth to lemurs_core crate
