+++ lemurs-core/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "lemurs"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+env_container.path = "../env_container"
+utmpx.path = "../utmpx"
+
+log.workspace = true
+nix.workspace = true
+libc.workspace = true
+users.workspace = true
+
+# Interacting with the kernel interfaces
+rand = "0.8.4"
+
+# Authentication and Fetching User Data
+pam = "0.7.0"
+pgs-files = "0.0.7"
+
+# Once Cell
+once_cell = "1.17.1"
