+++ lemurs-tui/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lemurs"
+name = "lemurs-tui"
 authors = ["Gijs Burghoorn <me@gburghoorn.com>"]
 description = "A TUI Display/Login Manager"
 version = "0.3.1"
@@ -17,32 +17,17 @@ license = "MIT or APACHE"
 
 [dependencies]
 chvt.path = "../chvt"
-env_container.path = "../env_container"
-utmpx.path = "../utmpx"
+lemurs_core = { package = "lemurs", path = "../lemurs-core" }
 
 # UI and TUI interface
 tui = { version = "0.16.0", features = [ "crossterm" ] }
 unicode-width = "0.1"
 crossterm = { version = "0.22" }
 
-# Interacting with the kernel interfaces
-rand = "0.8.4"
-nix.workspace = true
-
-# UTMPX
-libc.workspace = true
-
-# Authentication and Fetching User Data
-pam = "0.7.0"
-pgs-files = "0.0.7"
-users = "0.11.0"
-
-# Once Cell
-once_cell = "1.17.1"
-
 # Logging
 env_logger = { version = "0.9.0", default-features = false }
 log.workspace = true
+users.workspace = true
 
 # Configuration File Parsing
 toml = "0.5"
