+++ lemurs-tui/src/main.rs
@@ -12,49 +12,16 @@ use log::{error, info, warn};
 use tui::backend::CrosstermBackend;
 use tui::Terminal;
 
-mod auth;
 mod cli;
 mod config;
 mod info_caching;
-mod post_login;
 mod ui;
 
-use auth::try_auth;
 use config::Config;
-use post_login::{EnvironmentStartError, PostLoginEnvironment};
-
 use crate::cli::{Cli, Commands};
-
-use env_container::EnvironmentContainer;
-use utmpx::add_utmpx_entry;
-
-use self::{
-    auth::AuthenticationError,
-    config::ShellLoginFlag,
-    post_login::env_variables::{
-        set_basic_variables, set_display, set_seat_vars, set_session_params, set_session_vars,
-        set_xdg_common_paths,
-    },
-};
 
 const DEFAULT_CONFIG_PATH: &str = "/etc/lemurs/config.toml";
 const PREVIEW_LOG_PATH: &str = "lemurs.log";
-
-#[derive(Debug, Clone)]
-pub struct LemursConfig {
-    do_log: bool,
-    shell_login_flag: ShellLoginFlag,
-
-    client_log_path: String,
-    xserver_log_path: String,
-
-    xserver_timeout_secs: u16,
-    pam_service: String,
-
-    tty: u8,
-
-    x11_display: String,
-}
 
 fn merge_in_configuration(config: &mut Config, config_path: Option<&Path>) {
     let load_config_path = config_path.unwrap_or_else(|| Path::new(DEFAULT_CONFIG_PATH));
@@ -113,7 +80,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     if let Some(cmd) = cli.command {
         match cmd {
             Commands::Envs => {
-                let envs = post_login::get_envs(config.environment_switcher.include_tty_shell);
+                let envs = lemurs_core::post_login::get_envs(config.environment_switcher.include_tty_shell);
 
                 for (env_name, _) in envs.into_iter() {
                     println!("{env_name}");
@@ -212,112 +179,6 @@ pub fn tui_disable(mut terminal: Terminal<CrosstermBackend<io::Stdout>>) -> io::
     terminal.show_cursor()?;
 
     info!("Reset terminal environment");
-
-    Ok(())
-}
-
-struct Hooks<'a> {
-    pre_validate: Option<&'a dyn Fn()>,
-    pre_auth: Option<&'a dyn Fn()>,
-    pre_environment: Option<&'a dyn Fn()>,
-    pre_wait: Option<&'a dyn Fn()>,
-    pre_return: Option<&'a dyn Fn()>,
-}
-
-pub enum StartSessionError {
-    AuthenticationError(AuthenticationError),
-    EnvironmentStartError(EnvironmentStartError),
-}
-
-impl From<EnvironmentStartError> for StartSessionError {
-    fn from(value: EnvironmentStartError) -> Self {
-        Self::EnvironmentStartError(value)
-    }
-}
-
-impl From<AuthenticationError> for StartSessionError {
-    fn from(value: AuthenticationError) -> Self {
-        Self::AuthenticationError(value)
-    }
-}
-
-fn start_session(
-    username: &str,
-    password: &str,
-    post_login_env: &PostLoginEnvironment,
-    hooks: &Hooks<'_>,
-    config: &Config,
-) -> Result<(), StartSessionError> {
-    info!(
-        "Starting new session for '{}' in environment '{:?}'",
-        username, post_login_env
-    );
-
-    let config = LemursConfig {
-        do_log: config.do_log,
-        shell_login_flag: config.shell_login_flag,
-        client_log_path: config.client_log_path.clone(),
-        xserver_log_path: config.xserver_log_path.clone(),
-        xserver_timeout_secs: config.xserver_timeout_secs,
-        pam_service: config.pam_service.clone(),
-        tty: config.tty,
-        x11_display: config.x11_display.clone(),
-    };
-
-    if let Some(pre_validate_hook) = hooks.pre_validate {
-        pre_validate_hook();
-    }
-
-    let mut process_env = EnvironmentContainer::take_snapshot();
-
-    if let Some(pre_auth_hook) = hooks.pre_auth {
-        pre_auth_hook();
-    }
-
-    if matches!(post_login_env, PostLoginEnvironment::X { .. }) {
-        set_display(&config.x11_display, &mut process_env);
-    }
-    set_session_params(&mut process_env, post_login_env);
-
-    let auth_session = try_auth(username, password, &config.pam_service)?;
-
-    if let Some(pre_environment_hook) = hooks.pre_environment {
-        pre_environment_hook();
-    }
-
-    let tty = config.tty;
-    let uid = auth_session.uid;
-    let homedir = &auth_session.dir;
-    let shell = &auth_session.shell;
-
-    set_seat_vars(&mut process_env, tty);
-    set_session_vars(&mut process_env, uid);
-    set_basic_variables(&mut process_env, username, homedir, shell);
-    set_xdg_common_paths(&mut process_env, homedir);
-
-    let spawned_environment = post_login_env.spawn(&auth_session, &mut process_env, &config)?;
-
-    let pid = spawned_environment.pid();
-
-    let utmpx_session = add_utmpx_entry(username, tty, pid);
-    drop(process_env);
-
-    info!("Waiting for environment to terminate");
-
-    if let Some(pre_wait_hook) = hooks.pre_wait {
-        pre_wait_hook();
-    }
-
-    spawned_environment.wait();
-
-    info!("Environment terminated. Returning to Lemurs...");
-
-    if let Some(pre_return_hook) = hooks.pre_return {
-        pre_return_hook();
-    }
-
-    drop(utmpx_session);
-    drop(auth_session);
 
     Ok(())
 }
